+++ WndDesign/geometry/clamp.h
@@ -7,14 +7,14 @@
 BEGIN_NAMESPACE(WndDesign)
 
 
-constexpr int clamp(int number, Interval range) {
+constexpr float clamp(float number, Interval range) {
 	if (number < range.left()) { return range.left(); }
 	if (number >= range.right()) { return range.right() - 1; }
 	return number;
 }
 
-constexpr int clamp(Interval interval, Interval range) {
-	return clamp(clamp(interval.right(), range) - (int)interval.length, range);
+constexpr float clamp(Interval interval, Interval range) {
+	return clamp(clamp(interval.right(), range) - interval.length, range);
 }
 
 constexpr Point clamp(Point point, Rect region) {
@@ -22,7 +22,7 @@ constexpr Point clamp(Point point, Rect region) {
 }
 
 constexpr Point clamp(Rect rect, Rect region) {
-	return clamp(clamp(rect.RightBottom(), region) - Vector((int)rect.size.width, (int)rect.size.height), region);
+	return clamp(clamp(rect.RightBottom(), region) - Vector(rect.size.width, rect.size.height), region);
 }
