+++ WndDesign/wrapper/Border.h
@@ -17,12 +17,12 @@ public:
 protected:
 	struct Style {
 	public:
-		uint _width = 0;
-		uint _radius = 0;
+		float _width = 0.0f;
+		float _radius = 0.0f;
 		Color _color = color_transparent;
 	public:
-		constexpr Style& width(uint width) { _width = width; return *this; }
-		constexpr Style& radius(uint radius) { _radius = radius; return *this; }
+		constexpr Style& width(float width) { _width = width; return *this; }
+		constexpr Style& radius(float radius) { _radius = radius; return *this; }
 		constexpr Style& color(Color color) { _color = color; return *this; }
 	}border;
 
@@ -34,8 +34,8 @@ protected:
 	Vector GetInnerOffset() const { return Vector(border._width, border._width); }
 	Rect GetInnerRegion() const { return Rect(Point(border._width, border._width), inner_size); }
 protected:
-	virtual void OnSizeRefUpdate(Size size_ref) override { Wnd::OnSizeRefUpdate(Extend(size_ref, -(int)border._width)); }
-	virtual Size GetSize() override { return size = Extend(inner_size = Wnd::GetSize(), (int)border._width); }
+	virtual void OnSizeRefUpdate(Size size_ref) override { Wnd::OnSizeRefUpdate(Extend(size_ref, -border._width)); }
+	virtual Size GetSize() override { return size = Extend(inner_size = Wnd::GetSize(), border._width); }
 protected:
 	virtual ref_ptr<WndObject> HitTest(Point& point) override {
 		if (PointInRoundedRectangle(point, GetInnerRegion(), border._radius)) {
@@ -50,11 +50,11 @@ protected:
 protected:
 	virtual void OnDraw(FigureQueue& figure_queue, Rect draw_region) override {
 		figure_queue.Offset(GetInnerOffset(), [&]() { Wnd::OnDraw(figure_queue, draw_region - GetInnerOffset()); });
-		if (border._width > 0 && border._color.IsVisible()) {
-			if (border._radius > 0) {
-				figure_queue.add(point_zero, new RoundedRectangle(size, border._radius, (float)border._width, border._color));
+		if (border._width > 0.0f && border._color.IsVisible()) {
+			if (border._radius > 0.0f) {
+				figure_queue.add(point_zero, new RoundedRectangle(size, border._radius, border._width, border._color));
 			} else {
-				figure_queue.add(point_zero, new Rectangle(size, (float)border._width, border._color));
+				figure_queue.add(point_zero, new Rectangle(size, border._width, border._color));
 			}
 		}
 	}
